refactor: extract helper
